fix: add a fallback for root file/fullFile
This addresses an issue where the ouput from a test run with Cypress is not the
same as the test run with mocha directly. Specifically mocha does not set the
`file` property on the root suite and instead sets it on the sub suites or
tests whereas Cypress does the opposite.